Add the Substitute Edit.
