+++ edit/editor_test.go
@@ -9,7 +9,6 @@ import (
 	"testing"
 
 	"github.com/eaburns/T/edit/runes"
-	"github.com/eaburns/T/re1"
 )
 
 // String returns a string containing the entire editor contents.
@@ -71,148 +70,6 @@ func TestWhere(t *testing.T) {
 		if at != test.at || err != nil {
 			t.Errorf("ed.Where(%q)=%v,%v, want %v,<nil>", test.a, at, err, test.at)
 		}
-	}
-}
-
-func TestSubstituteFrom(t *testing.T) {
-	tests := []subTest{
-		{
-			init: "Hello, 世界!",
-			addr: ",", re: ".*", sub: ``, g: true, n: 1,
-			want: "", dot: addr{0, 0},
-		},
-		{
-			init: "Hello, 世界!",
-			addr: ",", re: "世界", sub: `World`, n: 1,
-			want: "Hello, World!", dot: addr{0, 13},
-		},
-		{
-			init: "Hello, 世界!",
-			addr: ",", re: "(.)", sub: `\1-`, g: true, n: 1,
-			want: "H-e-l-l-o-,- -世-界-!-", dot: addr{0, 20},
-		},
-		{
-			init: "abcabc",
-			addr: ",", re: "abc", sub: "defg", n: 1,
-			want: "defgabc", dot: addr{0, 7},
-		},
-		{
-			init: "abcabcabc",
-			addr: ",", re: "abc", sub: "defg", g: true, n: 1,
-			want: "defgdefgdefg", dot: addr{0, 12},
-		},
-		{
-			init: "abcabcabc",
-			addr: "/abcabc/", re: "abc", sub: "defg", g: true, n: 1,
-			want: "defgdefgabc", dot: addr{0, 8},
-		},
-		{
-			init: "abc abc",
-			addr: ",", re: "abc", sub: "defg", n: 1,
-			want: "defg abc", dot: addr{0, 8},
-		},
-		{
-			init: "abc abc",
-			addr: ",", re: "abc", sub: "defg", g: true, n: 1,
-			want: "defg defg", dot: addr{0, 9},
-		},
-		{
-			init: "abc abc abc",
-			addr: "/abc abc/", re: "abc", sub: "defg", g: true, n: 1,
-			want: "defg defg abc", dot: addr{0, 9},
-		},
-		{
-			init: "abcabc",
-			addr: ",", re: "abc", sub: "de", n: 1,
-			want: "deabc", dot: addr{0, 5},
-		},
-		{
-			init: "abcabcabc",
-			addr: ",", re: "abc", sub: "de", g: true, n: 1,
-			want: "dedede", dot: addr{0, 6},
-		},
-		{
-			init: "abcabcabc",
-			addr: "/abcabc/", re: "abc", sub: "de", g: true, n: 1,
-			want: "dedeabc", dot: addr{0, 4},
-		},
-		{
-			init: "func f()",
-			addr: ",", re: `func (.*)\(\)`, sub: `func (T) \1()`, g: true, n: 1,
-			want: "func (T) f()", dot: addr{0, 12},
-		},
-		{
-			init: "abcdefghi",
-			addr: ",", re: "(abc)(def)(ghi)", sub: `\0 \3 \2 \1`, n: 1,
-			want: "abcdefghi ghi def abc", dot: addr{0, 21},
-		},
-		{
-			init: "abc",
-			addr: ",", re: "abc", sub: `\1`, n: 1,
-			want: "", dot: addr{0, 0},
-		},
-		{
-			init: "abcabcabc",
-			addr: ",", re: "abc", sub: "def", n: 2,
-			want: "abcdefabc", dot: addr{0, 9},
-		},
-		{
-			init: "abcabcabc",
-			addr: ",", re: "abc", sub: "def", g: true, n: 2,
-			want: "abcdefdef", dot: addr{0, 9},
-		},
-		{
-			init: "abcabcabc",
-			addr: ",", re: "abc", sub: "def", g: false, n: 0,
-			want: "defabcabc", dot: addr{0, 9},
-		},
-		{
-			init: "abcabcabc",
-			addr: ",", re: "notpresent", sub: "def", g: false, n: 4,
-			want: "abcabcabc", dot: addr{0, 9},
-		},
-		{
-			init: "abcabcabc",
-			addr: ",", re: "abc", sub: "def", g: false, n: 4,
-			want: "abcabcabc", dot: addr{0, 9},
-		},
-	}
-	for _, test := range tests {
-		test.run(t)
-	}
-}
-
-type subTest struct {
-	init, addr, re, sub, want string
-	g                         bool
-	n                         int
-	dot                       addr
-}
-
-func (test subTest) run(t *testing.T) {
-	ed := NewEditor(NewBuffer())
-	defer ed.Close()
-	defer ed.buf.Close()
-	if err := ed.change(All, test.init); err != nil {
-		t.Fatalf("%#v, init failed", test)
-	}
-	addr, _, err := Addr([]rune(test.addr))
-	if err != nil {
-		t.Fatalf("%#v: bad address: %q", test, test.addr)
-	}
-	re, err := re1.Compile([]rune(test.re), re1.Options{})
-	if err != nil {
-		t.Fatalf("%#v: bad re: %q %v", test, test.re, err)
-	}
-	if err := ed.SubstituteFrom(addr, re, []rune(test.sub), test.g, test.n); err != nil {
-		t.Fatalf("%#v: ed.SubstituteFrom(%q, %q, %q, %v, %v)=%v, want nil",
-			test, test.addr, test.re, test.sub, test.g, test.n, err)
-	}
-	if ed.marks['.'] != test.dot {
-		t.Errorf("%#v: dot=%v, want %v\n", test, ed.marks['.'], test.dot)
-	}
-	if s := ed.String(); s != test.want {
-		t.Errorf("ed.String()=%q, want %q,nil\n", s, test.want)
 	}
 }
